requirements.txt
pillow
flask
torch